requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.2.0
billiard==4.2.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
datetime-truncate==1.1.1
Django==5.0.6
django-admin-soft-dashboard==1.0.12
django-celery-beat==2.6.0
django-cleanup==8.1.0
django-cron==0.6.0
django-js-routes==0.2.0
django-simple-history==3.7.0
django-timezone-field==6.1.0
django-vite==3.0.4
djangorestframework==3.15.2
fonttools==4.53.1
frozenlist==1.4.1
gunicorn==22.0.0
html5lib==1.1
idna==3.7
inertia-django==0.6.0
kombu==5.3.7
lxml==5.3.0
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
mysqlclient==2.2.4
numpy==2.0.1
oauthlib==3.2.2
oscrypto==1.3.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
prompt_toolkit==3.0.47
pycparser==2.22
pydyf==0.11.0
PyJWT==2.8.0
pypdf==4.3.1
pyphen==0.16.0
pypng==0.20220715.0
pytesseract==0.3.13
python-bidi==0.6.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.8
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
str2bool==1.1
text-unidecode==1.3
tinycss2==1.3.0
twilio==9.2.2
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.7.0
yarl==1.9.4
zopfli==0.2.3
